modified:   README.md 	modified:   inc/Symbols.h 	modified:   inc/console.h 	modified:   src/Symbols.c 	modified:   src/console.c 	modified:   src/main.c I have made a large change to main, where console-only operation is currently allowed. The data superstructure is now in use and funcs are being repurposed to use it if they are high enough in the chain.
Command list is being populated. Decided to use jump table and sum
hash. Func call in jump table (not func ptr) with argv and data so
that handling of command args takes place.

Only documentation right now is that which is printed to stderr in
the case of bad commands.

Also added more robust memory management, though this is still a
work in progress of course.